requirements.txt
altgraph==0.17.3
colorama==0.4.6
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pypng==0.20220715.0
pywin32-ctypes==0.2.0
qrcode==7.4.2
typing_extensions==4.5.0
